New zsh config.
Threw out most of the old zsh config in favour of a configuration
heavily based on grml zsh. It has a number of small modifications and
was split up into separate files. The largest change was in the prompt
code where it threw out the zsh themes and just did it.